code/Libraries/RZVisuals/pyproject.toml
[project]
name = "RZVisuals"
version = "0.1.0"
authors = [
  { name="Peter Preinesberger", email="peter.preinesberger@study.thws.de" },
]
description = "3D Rendering capabilities for visualizing trajectories of rodents"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux"
]
license-files = ["LICEN[CS]E*"]
dependencies = [
  "numpy",
  "pygame==2.6.1",
  "pyglm==2.8.2",
  "PyOpenGL==3.1.10"
]

[project.urls]
Homepage = "https://github.com/thws-mai/PM25_Mice"
Issues = "https://github.com/thws-mai/PM25_Mice/issues"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"
